requirement.txt
asgiref==3.2.7
backcall==0.1.0
colorama==0.4.3
decorator==4.4.2
Django==3.0.6
djangorestframework==3.11.0
Faker==4.1.0
ipdb==0.13.2
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
parso==0.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
traitlets==4.3.3
wcwidth==0.1.9
